Address PR #928 review feedback on ForeFlight export
